[Fix] 브로커 기본이미지일 때 기본이미지 변경 시 오류 수정 (#143)
* dev 최신화

* Feat : S3를 통한 이미지 업로드

* Fix : 기본 이미지 변경 오류 수정

+++ src/main/java/com/seahere/backend/s3/controller/S3Controller.java
@@ -35,11 +35,14 @@ public class S3Controller {
     @DeleteMapping("/delete")
     public ResponseEntity<Void> deleteFile(@AuthenticationPrincipal CustomUserDetails customUserDetails) {
         UserEntity user = s3Service.getUserById(customUserDetails.getUser().getUserId());
-        String fileName = user.getProfileImage();
-        s3Service.deleteFile(fileName);
-        user.updateProfileImage(null);
-        s3Service.saveUser(user);
-        return ResponseEntity.noContent().build();
+        if(user.getProfileImage()!= null){
+            String fileName = user.getProfileImage();
+            s3Service.deleteFile(fileName);
+            user.updateProfileImage(null);
+            s3Service.saveUser(user);
+            return ResponseEntity.noContent().build();
+        } else return ResponseEntity.noContent().build();
+
     }
 
     @GetMapping("/profile-image-url")
